requirements.txt
langchain
langchain-community
langchain_groq
langchain_huggingface
python-docx
weasyprint
cairosvg
markdown
PyPDF2
PyMuPDF
faiss-cpu
pypdf
langchain_openai
docx2txt
streamlit
PyMuPDF
langchain_community
python-docx
weasyprint